refactor: 이미지 관련 handler 함수 커스텀훅화
